Log In and Log Out works properly. Invalid log in attempts are blocked. If a user tries to access a certain page without logging in first, they will be redirected to the login page. Started working on Database Schemas for Event Admins and Judges. Since both Event Admins and Judges are Users, their models will inherit from the User Model.
Event Admins will have complete access to all the events that they created. They will be able to make changes to the event details, which entities are participating, etc. Event Admins can't judge unless they make themselves a Judge.

Judges will have limited access to the events that they are judging. They won't be able to make changes to the Event details or the participants, but they are the only ones that can access the scores of the participants. Each Judge is only responsible for their particular scores. They won't be able to make changes to other Judges' scores.